% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{FilterVariable}
\alias{FilterVariable}
\title{\code{FilterVariable}}
\usage{
FilterVariable(variable, combobox)
}
\arguments{
\item{variable}{A \code{\link{character}} (Text) or \code{\link{factor}} (Ordinal, Nominal) variable.}

\item{combobox}{A \code{\link{character}} vector containing selected options.}
}
\value{
A \code{\link{logical}} vector of length \code{length(variable)}.
}
\description{
Creates a filter variable.
}
